backend/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-Cors==6.0.1
langchain==0.3.10
langchain-openai==0.2.11
langchain-anthropic
langchain-google-genai
langchain-community
faiss-cpu==1.9.0.post1
pypdf==5.1.0
python-dotenv==1.0.1
sentence-transformers
torch




